test: Name 검증 테스트 작성
