+++ src/test/java/domain/CarTest.java
@@ -8,23 +8,6 @@ import static org.assertj.core.api.Assertions.*;
 
 class CarTest {
 
-    @DisplayName("자동차 이름의 길이가 1~5글자가 아니거나, null 이라면 예외를 던진다")
-    @ParameterizedTest
-    @CsvSource(value = {"''", "abcdef", "abcdefgh", "null"}, nullValues = "null")
-    void should_ThrowIllegalArgumentException_When_NameIsInvalid(String name) {
-        assertThatThrownBy(() -> new Car(name))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("[ERROR]");
-    }
-
-    @DisplayName("자동차 이름의 길이가 1~5글자라면 예외를 던지지 않는다")
-    @ParameterizedTest
-    @CsvSource({"a", "abc", "abcde"})
-    void should_DoesNotThrowException_When_NameIsValid(String name) {
-        assertThatCode(() -> new Car(name))
-                .doesNotThrowAnyException();
-    }
-
     @DisplayName("0 ~ 3의 값일때는 움직이지 않는다.")
     @ParameterizedTest
     @CsvSource({"0, 0", "1, 0", "2, 0", "3, 0"})
